backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
python-telegram-bot==13.7
django-cors-headers==3.10.0